% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinspacy.R
\name{bind_clinspacy}
\alias{bind_clinspacy}
\title{This function binds columns containing concept unique identifiers with which scispacy has
99% confidence of being present with values containing frequencies. Negated concepts,
as identified by negspacy's NegEx implementation, are ignored and do not count towards
the frequencies.}
\usage{
bind_clinspacy(df, text)
}
\arguments{
\item{df}{A data frame.}

\item{text}{A character string containing the name of the column to process.}
}
\value{
A data frame containing the original data frame as well as additional column names
for each UMLS concept unique identifer found with values containing frequencies.
}
\description{
This function binds columns containing concept unique identifiers with which scispacy has
99% confidence of being present with values containing frequencies. Negated concepts,
as identified by negspacy's NegEx implementation, are ignored and do not count towards
the frequencies.
}
\examples{
data(mtsamples)
mtsamples_with_cuis = bind_clinspacy(mtsamples[1:5,], text = 'description')
str(mtsamples_with_cuis)
}
