% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinspacy.R
\name{bind_clinspacy_embeddings}
\alias{bind_clinspacy_embeddings}
\title{This function binds columns containing concept embeddings for concepts with which scispacy has
99 percent confidence of being present with values containing frequencies. Negated concepts,
as identified by negspacy's NegEx implementation, are ignored and do not count towards
the embeddings. The concept embeddings are derived from the cui2vec_embeddings dataset
included with this package.}
\usage{
bind_clinspacy_embeddings(df, text, num_embeddings = 500, ...)
}
\arguments{
\item{df}{A data frame.}

\item{text}{A character string containing the name of the column to process.}

\item{num_embeddings}{The number of embeddings to return (must be a number 1 through 500).}

\item{...}{Arguments passed down to \code{\link{clinspacy}}}
}
\value{
A data frame containing the original data frame as well as additional column names
for each UMLS concept unique identifer found with values containing frequencies.
}
\description{
The embeddings are derived from Andrew Beam's
\href{https://github.com/beamandrew/cui2vec}{cui2vec R package}.
}
\details{
Citation

Beam, A.L., Kompa, B., Schmaltz, A., Fried, I., Griffin, W, Palmer, N.P., Shi, X.,
Cai, T., and Kohane, I.S.,, 2019. Clinical Concept Embeddings Learned from Massive
Sources of Multimodal Medical Data. arXiv preprint arXiv:1804.01486.

License

This data is made available under a
\href{https://creativecommons.org/licenses/by/4.0/}{CC BY 4.0 license}. The only change
made to the original dataset is the renaming of columns.
}
\examples{
data(mtsamples)
mtsamples_with_cuis = bind_clinspacy(mtsamples[1:5,], text = 'description')
str(mtsamples_with_cuis)
}
