% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinspacy.R
\name{clinspacy}
\alias{clinspacy}
\title{Performs biomedical named entity recognition, Unified Medical Language System (UMLS)
concept mapping, and negation detection using the Python spaCy, scispacy, and negspacy packages.}
\usage{
clinspacy(text)
}
\arguments{
\item{text}{A character string containing medical text that you would like to process.}
}
\value{
A data frame containing the UMLS concept unique identifiers (cui), entities,
lemmatized entities, and NegEx negation status (\code{TRUE} means negated, \code{FALSE} means *not* negated).
}
\description{
Performs biomedical named entity recognition, Unified Medical Language System (UMLS)
concept mapping, and negation detection using the Python spaCy, scispacy, and negspacy packages.
}
\examples{
clinspacy('This patient has diabetes and CKD stage 3 but no HTN.')
}
