% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cui2vec_data.R
\name{dataset_cui2vec_definitions}
\alias{dataset_cui2vec_definitions}
\title{Cui2vec concept definitions}
\format{
A data frame with 3053795 rows and 3 variables: \describe{
  \item{cui}{A Unified Medical Language System (UMLS) Concept Unique
  Identifier (CUI)} \item{semantic_type}{Semantic type of the CUI}
  \item{definition}{Definition of the CUI} }
}
\source{
\url{https://github.com/beamandrew/cui2vec}
}
\usage{
dataset_cui2vec_definitions()
}
\value{
Returns the cui2vec UMLS definitions as a data frame.
}
\description{
This dataset contains definitions for the Unified Medical Language System
(UMLS) Concept Unique Identifiers (CUIs). These come from Andrew Beam's
\href{https://github.com/beamandrew/cui2vec}{cui2vec R package}.
}
\details{
License

This data is made available under a
\href{https://github.com/beamandrew/cui2vec/blob/master/LICENSE.md}{MIT
license}. The data is copyrighted in 2019 by Benjamin Kompa, Andrew Beam, and
Allen Schmaltz. The only change made to the original dataset is the renaming
of columns.
}
